+++ java/client/src/main/java/glide/api/BaseClient.java
@@ -36,6 +36,7 @@ import static redis_request.RedisRequestOuterClass.RequestType.PExpire;
 import static redis_request.RedisRequestOuterClass.RequestType.PExpireAt;
 import static redis_request.RedisRequestOuterClass.RequestType.PfAdd;
 import static redis_request.RedisRequestOuterClass.RequestType.PfCount;
+import static redis_request.RedisRequestOuterClass.RequestType.PfMerge;
 import static redis_request.RedisRequestOuterClass.RequestType.RPop;
 import static redis_request.RedisRequestOuterClass.RequestType.RPush;
 import static redis_request.RedisRequestOuterClass.RequestType.SAdd;
@@ -600,5 +601,11 @@ public abstract class BaseClient
     @Override
     public CompletableFuture<Long> pfcount(@NonNull String[] keys) {
         return commandManager.submitNewCommand(PfCount, keys, this::handleLongResponse);
+    }
+
+    @Override
+    public CompletableFuture<String> pfmerge(String destKey, String[] sourceKeys) {
+        String[] arguments = ArrayUtils.addFirst(sourceKeys, destKey);
+        return commandManager.submitNewCommand(PfMerge, arguments, this::handleStringResponse);
     }
 }