+++ java/client/src/main/java/glide/api/commands/HyperLogLogBaseCommands.java
@@ -70,4 +70,26 @@ public interface HyperLogLogBaseCommands {
      * }</pre>
      */
     CompletableFuture<Long> pfcount(String[] keys);
+
+    /**
+     * Merges multiple HyperLogLog values into a unique value.
+     *
+     * <p>The computed merged HyperLogLog is created if it does not exist before (defaulting to an
+     * empty HyperLogLog).<br>
+     * If the destination variable exists, it is treated as one of the source HyperLogLog data set.
+     *
+     * @see <a href="https://redis.io/commands/pfmerge/">redis.io</a> for details.
+     * @param destKey The destination where to merge HyperLogLog data sets into.
+     * @param sourceKeys The source HyperLogLog to merge.
+     * @return <code>OK</code>
+     * @example
+     *     <pre>{@code
+     * String response = client.pfmerge("new_HLL", "old_HLL_1", "old_HLL_2").get();
+     * assert response.equals("OK"); // new HLL was created with merged content of old ones
+     *
+     * String response = client.pfmerge("old_HLL_1", "old_HLL_2", "old_HLL_3").get();
+     * assert response.equals("OK"); // content of existing HLLs was merged into existing variable
+     * }</pre>
+     */
+    CompletableFuture<String> pfmerge(String destKey, String[] sourceKeys);
 }